+++ integration_tests/src/main/python/mortgage_test.py
@@ -16,43 +16,6 @@ import pytest
 
 from asserts import assert_gpu_and_cpu_are_equal_iterator
 from marks import approximate_float, incompat, ignore_order, allow_non_gpu, limit
-from pyspark.sql.dataframe import DataFrame
-
-class MortgageRunner:
-  def __init__(self, mortgage_format, mortgage_acq_path, mortgage_perf_path):
-    self.mortgage_format = mortgage_format
-    self.mortgage_acq_path = mortgage_acq_path
-    self.mortgage_perf_path = mortgage_perf_path
-
-  def do_test_query(self, spark):
-    jvm_session = spark._jsparkSession
-    jvm = spark.sparkContext._jvm
-    acq = self.mortgage_acq_path
-    perf = self.mortgage_perf_path
-    run = jvm.com.nvidia.spark.rapids.tests.mortgage.Run
-    if self.mortgage_format == 'csv':
-        df = run.csv(jvm_session, perf, acq)
-    elif self.mortgage_format == 'parquet':
-        df = run.parquet(jvm_session, perf, acq)
-    elif self.mortgage_format == 'orc':
-        df = run.orc(jvm_session, perf, acq)
-    else:
-        raise AssertionError('Not Supported Format {}'.format(self.mortgage_format))
-
-    return DataFrame(df, spark.getActiveSession())
-
-@pytest.fixture(scope="session")
-def mortgage(request):
-    mortgage_format = request.config.getoption("mortgage_format")
-    mortgage_path = request.config.getoption("mortgage_path")
-    if mortgage_path is None:
-        std_path = request.config.getoption("std_input_path")
-        if std_path is None:
-            skip_unless_precommit_tests("Mortgage tests are not configured to run")
-        else:
-            yield MortgageRunner('parquet', std_path + '/parquet_acq', std_path + '/parquet_perf')
-    else:
-        yield MortgageRunner(mortgage_format, mortgage_path + '/acq', mortgage_path + '/perf')
 
 @incompat
 @approximate_float
