+++ integration_tests/src/main/python/spark_session.py
@@ -14,8 +14,8 @@
 
 import os
 from conftest import is_allowing_any_non_gpu, get_non_gpu_allowed, get_validate_execs_in_gpu_plan, is_databricks_runtime
-from spark_init_internal import get_spark_i_know_what_i_am_doing, spark_version
 from pyspark.sql import DataFrame
+from spark_init_internal import get_spark_i_know_what_i_am_doing, spark_version
 
 def _from_scala_map(scala_map):
     ret = {}
